requirements.txt
Pillow==1.7.8
South==0.7.6
distribute==0.6.27
dj.choices==0.9.0
django-bootstrap-toolkit==2.6.22
django-braces==0.2.2
django-follow==0.6.1
django-guardian==1.0.4
django-stream==0.2.3
fuzzywuzzy==0.1
geopy==0.94.2
pytz==2012j
selenium==2.28.0
wsgiref==0.1.2
https://www.djangoproject.com/download/1.5c1/tarball/#egg=Django
